Checkpoints configuration store tests.
